requirements.txt
ultralytics
opencv-python
numpy
python-dotenv
Flask
yt-dlp